requirements.txt
Flask
google-genai
reportlab
gunicorn
